Add load/save to CSV

+++ src/RegulusLibrary/DataStructures/BirdRecord.cs
@@ -5,8 +5,8 @@ public record BirdRecord
     public int Id { get; set; }
     public string? SpeciesCode { get; set; }
     public string? Sex { get; set; }
-    public double? Wing { get; set; }
-    public double? Tail { get; set; }
-    public double? Weight { get; set; }
-    public DateTime? DateTime{ get; set; }
+    public decimal? Wing { get; set; }
+    public decimal? Tail { get; set; }
+    public decimal? Weight { get; set; }
+    public DateTime? DateTime { get; set; }
 }
